feat: 글 업로드 기능

+++ src/main/java/com/example/demo/entity/Pictures.java
@@ -13,12 +13,12 @@ public class Pictures {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    private Integer id;
+    private Long id;
 
     @Column(name = "picture")
     private String picture;
 
     @Column(name = "post_id")
-    private Integer postId;
+    private Long postId;
 
 }
