+++ src/main/java/com/example/demo/entity/Posts.java
@@ -14,7 +14,7 @@ public class Posts {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    private Integer id;
+    private Long id;
 
     @Column(name = "title")
     private String title;
@@ -35,6 +35,9 @@ public class Posts {
     private LocalDateTime updatedAt;
 
     @Column(name = "author_Id")
-    private Integer authorId;
+    private Long authorId;
+
+//    @ManyToOne
+//    private Members member;
 
 }
